requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.11.4
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
PyJWT==2.10.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
